refactor: use outlet for sub route modal
