+++ src/App.tsx
@@ -1,13 +1,16 @@
 import { Provider as ReduxProvider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
 
 import Store from 'state/store';
 import Routes from 'routes';
 
 function App() {
   return (
-    <ReduxProvider store={Store}>
-      <Routes />
-    </ReduxProvider>
+    <BrowserRouter>
+      <ReduxProvider store={Store}>
+        <Routes />
+      </ReduxProvider>
+    </BrowserRouter>
   );
 }
